L1-218: Refector and cleanup in pallet-balances-maintenance.py
